New block header

+++ src/test/skiplist_tests.cpp
@@ -140,7 +140,7 @@ BOOST_AUTO_TEST_CASE(findearliestatleast_test) {
         }
     }
     // Check that we set nTimeMax up correctly.
-    unsigned int curTimeMax = 0;
+    int64_t curTimeMax = 0;
     for (size_t i = 0; i < vBlocksMain.size(); ++i) {
         curTimeMax = std::max(curTimeMax, vBlocksMain[i].nTime);
         BOOST_CHECK(curTimeMax == vBlocksMain[i].nTimeMax);
